virt-aa-helper-test: Fix failure due to the new disk format probing option
Commit 68719c4bddb85fbcc931a5b7d99ac7c8a0af09b0 added the disk format
probing option. This makes virt-aa-helper-test fail because the domain
config didn't specifiy the disk format and it didn't pass '-p 1' to
virt-aa-helper to allow disk format probing.

Specify the disk format in the domain config. Pass the '-p 1' option
to virt-aa-helper for the test case with two disks. This way this test
also covers this new option.